// src/utils/enum.ts
export enum COLORS {
  BLACK = "#000000",
  PRIMARY = "#FFD94A",
  BLUE_OPACITY = "#F0EFFF",
  ICON_COLOR = "#6C5FFC",
  TEXT_COLOR = "#878AA6",
  GREEN_COLOR = "#7AA93C",
  TABLE_HEADER_COLOR = "#F8F9FD",
  RED_OPACITY = "#ff000020",
  RED = "#ff0000",
  GREEN_OPACITY = "#00800020",
  TRANSPARENT = "transparent",
  WHITE = "#ffffff",
  BODYCOLOR = "#131313",
}

export enum ATTENDANCE_STATUS {
  ABSENT = "Absent",
  PRESENT = "Present",
}

export enum USER_STATUS {
  ACTIVE = "Active",
  INACTIVE = "Inactive",
}